Move custom settings to a single block
This commit moves all custom settings to a single block. It is meant
to keep differences between this branch and upstream grml-etc-core to
a minimum and to make future merges from master easier.

Local customization could also be moved to the user-specific local
configuration file (~/.tmux.conf.local), but keeping it in here makes
it easier to drop this tmux.conf into /etc on a new system.